fix: refactor architecture and remove unnecessary checks
